btrfs: use last-allocated extent prior to offset to determine alloc hint
get_extent_allocation_hint currently punts when it encounters
a hole in the file and reverts to using the first offset within
the file as a hint.  When using a sparse file that is filled in
randomly, this can result in pathological fragmentation as each
write is fulfilled on disk with the next available extent.

This patch changes the search to locate the requested offset and then
search backward for the last mapped offset.  Once located, we can
provide a hint that reflects where the requested offset would land
if it were allocated linearly.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>