btrfs: fix incorrect offset for searching in find_free_extent
Since commit 963030817060e (Btrfs: use hybrid extents+bitmap rb tree
for free space), we've been passing the result of offset_to_bitmap()
to tree_search_offset.  Initially, this was to prevent spurious ENOSPC
issues if we were too deep into the bitmap to find the target offset.
These days, we'll back off and reset the hint before repeating the
search.

The fallout of passing the result of offset_to_bitmap() is that we
effectively round the hint down to the next bytes_per_bitmap boundary,
which can be up to 128MiB below the requested offset.  The result is
an allocation that falls before the requested one and potentially
before the previously allocated extent, which can create fragmentation
issues.

Since we back off and retry with a different hint, we can use the
correct offset in tree_search_offset.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>