btrfs: honor hint in find_free_extent
If we provide a hint that falls in the middle of a free extent, we still
choose the beginning of the extent for allocation.  This generally works
when the workload involves a typical growing file.  The hint will more or
less be ignored, but the extents will mostly be allocated sequentially.
When the workload is writing into a sparse file randomly, the layout on
disk will match the random workload.  For a qcow-style VM image, that can
have a disastrous effect on performance.

This patch splits the free space extent if the range starting at the
hint falls entirely within the free space extent. Otherwise, the existing
behavior of using the first extent in the free space extent is used.

The restoration of the free space prior to the offset within the free
space extent is already handled via the alignment mechanism.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>