btrfs: handle failure to re-add free space while allocating
If we're allocating in the middle of a free extent, either due to
hinting or alignment, we convert the free space entry to describe the
part of the free extent after the allocation and re-link it.  Then we
add the part before the allocation back to the cache separately.  This
implies allocating a new free space entry, which can fail.  We currently
ignore this failure silently, which will leak free space.

If we drop the lock and insert the new free entry before linking the
truncated old entry, we can handle the failure gracefully by using the
start of the free space extent as we did prior to the previous commit.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>